requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
openai==1.3.0
anthropic==0.18.0
pydantic==2.5.0
python-dotenv==1.0.0
pillow==10.1.0
pytesseract==0.3.10
google-api-python-client==2.100.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
reportlab==4.0.7
jinja2==3.1.2

